feat(TTS): 增加腾讯云TTS和WavPlayer

+++ robot/utils.py
@@ -1,4 +1,5 @@
 import os
+import tempfile
 
 def get_file_content(filePath):
     """ 读取文件 """
@@ -7,5 +8,11 @@ def get_file_content(filePath):
 
 def check_and_delete(fp):
     """ 检查并删除文件 """
-    if os.path.exists(fp):
+    if isinstance(fp, str) and os.path.exists(fp):
         os.remove(fp)
+
+def write_temp_file(data, suffix):
+    with tempfile.NamedTemporaryFile(suffix=suffix, delete=False) as f:
+        f.write(data)
+        tmpfile = f.name
+    return tmpfile
